Extract shared message serialization helpers
The conversations and requests servicers each had their own copy of
message_to_pb, and each was missing message types the other rendered:
conversations' copy didn't handle host_request_status_changed, and
requests' copy didn't handle the group chat control messages. Move a
single complete version (plus the host request status maps) to
couchers.helpers.messages.
